LayoutTests/fast/css/parse-justify-self-expected.txt

Test that setting and getting justify-self works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting justify-self set through CSS
PASS getComputedStyle(justifySelfBaseline, '').getPropertyValue('justify-self') is 'baseline'
PASS getComputedStyle(justifySelfStretch, '').getPropertyValue('justify-self') is 'stretch'
PASS getComputedStyle(justifySelfStart, '').getPropertyValue('justify-self') is 'start'
PASS getComputedStyle(justifySelfEnd, '').getPropertyValue('justify-self') is 'end'
PASS getComputedStyle(justifySelfCenter, '').getPropertyValue('justify-self') is 'center'
PASS getComputedStyle(justifySelfSelfEnd, '').getPropertyValue('justify-self') is 'self-end'
PASS getComputedStyle(justifySelfSelfStart, '').getPropertyValue('justify-self') is 'self-start'
PASS getComputedStyle(justifySelfLeft, '').getPropertyValue('justify-self') is 'left'
PASS getComputedStyle(justifySelfRight, '').getPropertyValue('justify-self') is 'right'
PASS getComputedStyle(justifySelfEndUnsafe, '').getPropertyValue('justify-self') is 'end unsafe'
PASS getComputedStyle(justifySelfCenterUnsafe, '').getPropertyValue('justify-self') is 'center unsafe'
PASS getComputedStyle(justifySelfSelfEndSafe, '').getPropertyValue('justify-self') is 'self-end safe'
PASS getComputedStyle(justifySelfSelfStartSafe, '').getPropertyValue('justify-self') is 'self-start safe'
PASS getComputedStyle(justifySelfRightSafe, '').getPropertyValue('justify-self') is 'right safe'
PASS getComputedStyle(justifySelfLeftUnsafe, '').getPropertyValue('justify-self') is 'left unsafe'

Test initial value of justify-self through JS
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'

Test getting and setting justify-self through JS
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'center'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start unsafe'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'

Test bad combinations of justify-self
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'

Test the value 'initial'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'center'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'start'

Test the value 'inherit'
PASS getComputedStyle(parentElement, '').getPropertyValue('justify-self') is 'end'
PASS getComputedStyle(element, '').getPropertyValue('justify-self') is 'end'
PASS successfullyParsed is true

TEST COMPLETE
